Hunting is now almost Bug Free, Some ints are now Doubles to help preven Data loss, and Death now functions correctly

+++ RPGame/Monsters.h
@@ -2,8 +2,8 @@
 class Monsters
 {
 public:
-	int maxHealthPoints;
-	int currentHealthPoints;
+	double maxHealthPoints;
+	double currentHealthPoints;
 	int attack;
 	int defence;
 	std::string name;
